Extract PaletteShape into separate file
